feat: Add parallax background, text reveal, and animated paint effects components; include particle colors and carousel images constants
