+++ src/constansts/carousel-images.tsx
@@ -1,0 +1,26 @@
+// Hero carousel images
+const carouselImages = [
+  {
+    src: "https://images.unsplash.com/photo-1600210492493-0946911123ea?q=80&w=1974&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D",
+    alt: "Painting Services",
+    title: "Crafting Beautiful Spaces with Precision and Passion",
+    description:
+      "We are a locally owned and managed painting contractor providing superior quality painting services for over 10 years.",
+  },
+  {
+    src: "https://images.unsplash.com/photo-1560184897-ae75f418493e?q=80&w=2070&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D",
+    alt: "Interior Painting",
+    title: "Transform Your Interior Spaces",
+    description:
+      "Our expert team brings new life to your home with premium paints and meticulous attention to detail.",
+  },
+  {
+    src: "https://images.unsplash.com/photo-1567954970774-58d6aa6c50dc?q=80&w=1932&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D",
+    alt: "Exterior Painting",
+    title: "Elevate Your Home's Curb Appeal",
+    description:
+      "Weather-resistant, long-lasting exterior painting solutions that protect and beautify your property.",
+  },
+];
+
+export default carouselImages;
