+++ src/constansts/particle-colors.tsx
@@ -1,0 +1,3 @@
+const particleColors = ["#FF5A5F", "#3498DB", "#2ECC71", "#F1C40F", "#9B59B6"];
+
+export default particleColors;
